Removed the subscription diagnotic settings file and joined the code with the subscriptions file, changed the recordings file to allow for less boilerplate on tests.
